Adiciona atributo 'User' em 'Employee'
